+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## [Unreleased]
 
+## 4.3.1 2020-10-20
+
+### Fixed
+
+- Avoid `fetchScannedHostDetails` step failure when no hosts found
+- Avoid `fetchScannedHostFindings` step failure when no hosts found, or when
+  details could not be loaded for hosts
+
 ## 4.3.0 2020-10-20
 
 ### Added